+++ example/java/config/task/completecondition/TestNoColumn.java
@@ -3,7 +3,7 @@ package config.task.completecondition;
 public final class TestNoColumn implements config.task.completecondition.Completecondition {
     @Override
     public config.task.Completeconditiontype type() {
-        return config.task.Completeconditiontype.TESTNOCOLUMN;
+        return config.task.Completeconditiontype.TEST_NO_COLUMN;
     }